+++ src/GenericModalLayout/GenericModalLayout.protractor.driver.js
@@ -1,0 +1,8 @@
+const GenericModalLayoutDriverFactory = component => {
+  return {
+    getElement: () => component,
+    isFullscreen: async () => await component.getAttribute('data-fullscreen') === 'true'
+  };
+};
+
+export default GenericModalLayoutDriverFactory;