+++ stories/TableActionCell/examples/OnlySecondaryExample.js
@@ -9,9 +9,9 @@ const Example = () => (
     <TableActionCell
       dataHook="story-only-secondary"
       secondaryActions={[
-        {text: 'Download', icon: <Download/>, onClick: () => window.alert('Download action was diviggered.')},
-        {text: 'Duplicate', icon: <Duplicate/>, onClick: () => window.alert('Duplicate action was diviggered.')},
-        {text: 'Print', icon: <Print/>, onClick: () => window.alert('Print action was diviggered.')}
+        {text: 'Download', icon: <Download/>, onClick: () => window.alert('Download action was triggered.')},
+        {text: 'Duplicate', icon: <Duplicate/>, onClick: () => window.alert('Duplicate action was triggered.')},
+        {text: 'Print', icon: <Print/>, onClick: () => window.alert('Print action was triggered.')}
       ]}
       numOfVisibleSecondaryActions={2}
       />