+++ stories/TableActionCell/examples/OnlyVisibleSecondaryExample.js
@@ -9,8 +9,8 @@ const Example = () => (
     <TableActionCell
       dataHook="story-only-visible-secondary"
       secondaryActions={[
-        {text: 'Download', icon: <Download/>, onClick: () => window.alert('Download action was diviggered.')},
-        {text: 'Duplicate', icon: <Duplicate/>, onClick: () => window.alert('Duplicate action was diviggered.')}
+        {text: 'Download', icon: <Download/>, onClick: () => window.alert('Download action was triggered.')},
+        {text: 'Duplicate', icon: <Duplicate/>, onClick: () => window.alert('Duplicate action was triggered.')}
       ]}
       numOfVisibleSecondaryActions={2}
       />