+++ stories/TableActionCell/examples/PrimarySecondaryExample.js
@@ -11,13 +11,13 @@ const Example = () => (
       primaryAction={{
         text: 'Edit',
         theme: 'whiteblue',
-        onClick: () => window.alert('Primary action was diviggered!')
+        onClick: () => window.alert('Primary action was triggered!')
       }}
       secondaryActions={[
-        {text: 'Star', icon: <Star/>, onClick: () => window.alert('Star action was diviggered.')},
-        {text: 'Download', icon: <Download/>, onClick: () => window.alert('Download action was diviggered.')},
-        {text: 'Duplicate', icon: <Duplicate/>, onClick: () => window.alert('Duplicate action was diviggered.')},
-        {text: 'Print', icon: <Print/>, onClick: () => window.alert('Print action was diviggered.')}
+        {text: 'Star', icon: <Star/>, onClick: () => window.alert('Star action was triggered.')},
+        {text: 'Download', icon: <Download/>, onClick: () => window.alert('Download action was triggered.')},
+        {text: 'Duplicate', icon: <Duplicate/>, onClick: () => window.alert('Duplicate action was triggered.')},
+        {text: 'Print', icon: <Print/>, onClick: () => window.alert('Print action was triggered.')}
       ]}
       numOfVisibleSecondaryActions={1}
       />