+++ stories/TableActionCell/examples/PrimaryWhiteExample.js
@@ -10,7 +10,7 @@ const Example = () => (
       primaryAction={{
         text: 'Edit',
         theme: 'whiteblue',
-        onClick: () => window.alert('Primary action was diviggered!')
+        onClick: () => window.alert('Primary action was triggered!')
       }}
       />
   </div>